+++ src/main.rs
@@ -11,10 +11,10 @@ mod app_state;
 mod cli;
 mod config;
 mod errors;
+mod layout;
 mod logging;
 mod model;
 mod mutator;
-mod render;
 mod shared;
 mod states;
 mod store;
